+++ tests/text_to_ast/text_only.rs
@@ -205,7 +205,7 @@ fn link() {
     }
 
     for input in &test_cases {
-        println!("testing {}", format!("<{}>", input));
+        println!("testing <{}>", input);
         assert_eq!(
             parse_only_text(input),
             vec![Link {